requirements.txt
absl-py==2.1.0
accelerate==0.33.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==9.0.2
backcall==0.2.0
bitsandbytes==0.41.0
black==21.4b2
blis==0.7.11
Brotli @ file:///croot/brotli-split_1714483155106/work
catalogue==2.0.10
certifi==2023.11.17
cffi==1.17.0
charset-normalizer==3.3.2
click==7.1.2
cloudpathlib==0.19.0
cloudpickle==3.0.0
cmake==3.28.1
comm==0.2.2
confection==0.1.5
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dataclasses-json==0.5.14
debugpy==1.8.5
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
-e git+https://github.com/yiwengxie/Chat-Video.git@cd6f83c37ceee0d03654dc964ff86fe87bdc4fc3#egg=detectron2
distro==1.9.0
docker-pycreds==0.4.0
easydict==1.13
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.0
executing==2.1.0
fairscale==0.4.13
fastapi==0.112.2
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock @ file:///croot/filelock_1700591183607/work
flash_attn==2.6.3
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.41
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
gradio==3.24.1
gradio_client==0.0.7
greenlet==3.0.3
-e git+https://github.com/yiwengxie/Chat-Video.git@cd6f83c37ceee0d03654dc964ff86fe87bdc4fc3#egg=groundingdino&subdirectory=projects/GroundedSAM2/grounding_dino
grpcio==1.65.5
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
hydra-core==1.3.2
idna==3.6
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.11.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.5.0
joblib==1.3.2
jpeg4py==0.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.0.101
langchain-community==0.0.38
langchain-core==0.2.39
langchain-openai==0.1.23
langchain-text-splitters==0.0.2
langcodes==3.4.0
langsmith==0.1.117
language_data==1.2.0
lazy_loader==0.4
librosa==0.9.2
linkify-it-py==2.0.2
lit==17.0.6
-e git+https://github.com/dvlab-research/LLaMA-VID.git@d1074f3662a772d1b3c723416af59314ba593f67#egg=llamavid
# Editable install with no version control (llava==1.7.0.dev0)
-e /data/junke_dont_remove/Chat-Video/projects/LLaVA-NeXT
llvmlite==0.43.0
lvis @ git+https://github.com/lvis-dataset/lvis-api.git@7d7f07def11da91f8b2710ce352c62a78fd5a7ad
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==2.2.0
markdown2==2.4.12
MarkupSafe==2.1.4
marshmallow==3.22.0
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
mmcv==2.2.0
mmengine==0.10.4
more-itertools==10.5.0
motmetrics==1.4.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
MultiScaleDeformableAttention==1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numba==0.60.0
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
open_clip_torch==2.26.1
openai==0.27.2
openai-whisper==20230314
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==23.2
pandas==2.2.0
parso==0.8.4
pathlib_abc==0.1.1
pathspec==0.12.1
pathy==0.11.0
peft==0.4.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.2
pooch==1.8.2
portalocker==2.10.1
preshed==3.0.9
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools @ git+https://github.com/XD7479/cocoapi.git@772b85a787ef554a732d4cb9abe317d61602af83#subdirectory=PythonAPI
pycparser==2.22
pydantic==1.8.2
pydantic_core==2.23.3
pydot==3.0.1
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.32.1
regex==2023.12.25
requests @ file:///croot/requests_1721410876868/work
resampy==0.4.3
rich==13.8.0
rpds-py==0.17.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
safetensors==0.4.4
-e git+https://github.com/yiwengxie/Chat-Video.git@cd6f83c37ceee0d03654dc964ff86fe87bdc4fc3#egg=SAM_2&subdirectory=projects/GroundedSAM2
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.12.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.39.2
setproctitle==1.3.3
Shapely==1.7.1
shortuuid==1.0.11
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
soundfile==0.12.1
spacy==3.0.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.53
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.4
supervision==0.22.0
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
thinc==8.0.17
threadpoolctl==3.2.0
tifffile==2024.8.10
tiktoken==0.7.0
tikzplotlib==0.10.1
timm==0.6.13
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
triton==2.0.0
typer==0.3.2
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2023.4
uc-micro-py==1.0.2
urllib3 @ file:///croot/urllib3_1718912636303/work
uvicorn==0.26.0
visdom==0.2.4
wasabi==0.10.1
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
whisper==1.1.10
xmltodict==0.13.0
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.20.1
